Big PR: Implement path finding, implement location-based retry actions. (#69)
* implement pathfinding and some cleanup

* added support for some requests to be re-tried after pathfinding to a non-adjacent location, i.e. use item that's far.

* more support for pathfinding on other actions

* added tile caching to make pathfinding work correctly, implemented liquid types and fixed the values sent to the client.

* address comments

+++ OpenTibia.Server/MovementEvents/BodyToContainerMovementEvent.cs
@@ -70,7 +70,7 @@ namespace OpenTibia.Server.MovementEvents
             {
                 bool moveSuccessful = thingMoving is IItem item &&
                                       creatureFinder.FindCreatureById(targetCreatureId) is IPlayer targetPlayer &&
-                                      this.Game.PerformItemMovement(item, targetPlayer.Inventory[fromCreatureSlot] as IContainerItem, targetPlayer.GetContainerById(toCreatureContainerId), 0, toCreatureContainerIndex, amount, this.Requestor);
+                                      this.Game.PerformItemMovement(item, targetPlayer.Inventory[(byte)fromCreatureSlot] as IContainerItem, targetPlayer.GetContainerById(toCreatureContainerId), 0, toCreatureContainerIndex, amount, this.Requestor);
 
                 if (!moveSuccessful)
                 {
